requirements.txt
boto3>=1.26.0
azure-storage-blob>=12.16.0
google-cloud-storage>=2.9.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0 